requirements.txt
# Core dependencies
pyautogui>=0.9.54
pillow>=10.0.0
pytesseract>=0.3.10

# AI and API dependencies
openai>=1.0.0
python-dotenv>=1.0.0

# Terminal UI
prompt_toolkit>=3.0.0
rich>=13.0.0

# Utility dependencies
pyyaml>=6.0.0
requests>=2.31.0
asyncio>=3.4.3

# Development dependencies
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0
mypy>=1.0.0
